Break out UTs into class-specific files
