+++ Anytime_RRT_Sharp/src/testdata/obstacles.txt
