+++ Anytime_RRT_Sharp/src/tests/UnitTests.cpp
@@ -1,0 +1,10 @@
+#include <gtest/gtest.h>
+#include "ARRTSTests.hpp"
+#include "ObstacleTests.hpp"
+#include "VehicleTests.hpp"
+
+int main(int argc, char* argv[])
+{
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
